# core/types, rlp: optimize derivesha