--- a/rlp/encode.go
+++ b/rlp/encode.go
@@ -604,17 +604,76 @@ func makeEncoderWriter(typ reflect.Type) writer {
 	return w
 }
 
-func PutInt(b []byte, i int) (size int) {
+// AppendInt appends the rlp-encoding of i to b, and returns the
+// resulting slice
+func AppendInt(b []byte, i uint64) []byte {
 	if i == 0 {
-		b[0] = 0x80
-		return 1
+		return append(b, 0x80)
 	} else if i < 128 {
-		b[0] = byte(i)
-		return 1
+		return append(b, byte(i))
 	}
-	s := putint(b[1:], uint64(i))
-	b[0] = 0x80 + byte(s)
-	return s + 1
+	switch {
+	case i < (1 << 8):
+		return append(b, 0x81, byte(i))
+	case i < (1 << 16):
+		return append(b, 0x82,
+			byte(i>>8),
+			byte(i),
+		)
+	case i < (1 << 24):
+		return append(b, 0x83,
+			byte(i>>16),
+			byte(i>>8),
+			byte(i),
+		)
+	case i < (1 << 32):
+		return append(b, 0x84,
+			byte(i>>24),
+			byte(i>>16),
+			byte(i>>8),
+			byte(i),
+		)
+	case i < (1 << 40):
+		return append(b, 0x85,
+			byte(i>>32),
+			byte(i>>24),
+			byte(i>>16),
+			byte(i>>8),
+			byte(i),
+		)
+
+	case i < (1 << 48):
+		return append(b, 0x86,
+			byte(i>>40),
+			byte(i>>32),
+			byte(i>>24),
+			byte(i>>16),
+			byte(i>>8),
+			byte(i),
+		)
+	case i < (1 << 56):
+		return append(b, 0x87,
+			byte(i>>48),
+			byte(i>>40),
+			byte(i>>32),
+			byte(i>>24),
+			byte(i>>16),
+			byte(i>>8),
+			byte(i),
+		)
+
+	default:
+		return append(b, 0x88,
+			byte(i>>56),
+			byte(i>>48),
+			byte(i>>40),
+			byte(i>>32),
+			byte(i>>24),
+			byte(i>>16),
+			byte(i>>8),
+			byte(i),
+		)
+	}
 }
 
 // putint writes i to the beginning of b in big endian byte