This PR contains a minor optimization in derivesha, by exposing the rlp int-encoding and making use of it to write integers directly to a buffer (an rlp integer is known to never require more than `9` bytes total). 
This idea (using `rlp.PutInt`) might be useful in other places too. Is it a good idea? Consider it a suggestion.  

The code assumes, just as before, that the `hasher` (a trie) will copy the key internally, which it does when doing `keybytesToHex(key)`. 


```
name                       old time/op    new time/op    delta
DeriveSha200/std_trie-6       590µs ± 8%     546µs ± 4%   -7.43%  (p=0.000 n=9+9)
DeriveSha200/stack_trie-6     492µs ± 2%     458µs ± 5%   -7.05%  (p=0.000 n=9+9)

name                       old alloc/op   new alloc/op   delta
DeriveSha200/std_trie-6       275kB ± 0%     274kB ± 0%   -0.61%  (p=0.000 n=10+10)
DeriveSha200/stack_trie-6    55.0kB ± 0%    52.5kB ± 0%   -4.52%  (p=0.000 n=10+9)

name                       old allocs/op  new allocs/op  delta
DeriveSha200/std_trie-6       2.87k ± 0%     2.67k ± 0%   -6.97%  (p=0.000 n=10+10)
DeriveSha200/stack_trie-6     1.27k ± 0%     1.07k ± 0%  -15.75%  (p=0.000 n=10+10)
```