@holiman 2020-10-30
Yes, you are totally right. I had the same thought a few days ago, will update it now