@holiman 2020-10-30
Fixed. The latest changes (compared to the first changes):
```
name                       old time/op    new time/op    delta
DeriveSha200/std_trie-6       546µs ± 4%     520µs ± 4%  -4.74%  (p=0.000 n=9+9)
DeriveSha200/stack_trie-6     458µs ± 5%     442µs ± 1%  -3.41%  (p=0.000 n=9+8)

name                       old alloc/op   new alloc/op   delta
DeriveSha200/std_trie-6       274kB ± 0%     274kB ± 0%  -0.02%  (p=0.001 n=10+10)
DeriveSha200/stack_trie-6    52.5kB ± 0%    52.5kB ± 0%    ~     (p=0.067 n=9+10)

name                       old allocs/op  new allocs/op  delta
DeriveSha200/std_trie-6       2.67k ± 0%     2.67k ± 0%    ~     (all equal)
DeriveSha200/stack_trie-6     1.07k ± 0%     1.07k ± 0%    ~     (all equal)

```
And compared to `master`: 
```
name                       old time/op    new time/op    delta
DeriveSha200/std_trie-6       590µs ± 8%     520µs ± 4%  -11.82%  (p=0.000 n=9+9)
DeriveSha200/stack_trie-6     492µs ± 2%     442µs ± 1%  -10.22%  (p=0.000 n=9+8)

name                       old alloc/op   new alloc/op   delta
DeriveSha200/std_trie-6       275kB ± 0%     274kB ± 0%   -0.63%  (p=0.000 n=10+10)
DeriveSha200/stack_trie-6    55.0kB ± 0%    52.5kB ± 0%   -4.53%  (p=0.000 n=10+10)

name                       old allocs/op  new allocs/op  delta
DeriveSha200/std_trie-6       2.87k ± 0%     2.67k ± 0%   -6.97%  (p=0.000 n=10+10)
DeriveSha200/stack_trie-6     1.27k ± 0%     1.07k ± 0%  -15.75%  (p=0.000 n=10+10)
```