@fjl 2020-11-04
BTW, I think an additional optimization might be possible: right now, we are encoding the elements to a new byte slice (in `GetRlp`). They could also be encoded into a reusable `bytes.Buffer`. It would need some changes to the DerivableList interface though.