@holiman 2020-11-04
I don't think that would work well, the trie needs exclusive access to the elements anyway. Right now, it just assumes it has, but if we reuse buffers for value, we'd have to copy on TryUpdate, and thus the optimization would be mostly moot(?).